more dtype fixes for half-precision training

+++ mb_vae_dti/training/diffusion/utils.py
@@ -212,9 +212,9 @@ def mask_distributions(true_X, true_E, pred_X, pred_E, node_mask):
     Originally defined in DiffMS/src/diffusion/diffusion_utils.py
     """
 
-    row_X = torch.zeros(true_X.size(-1), dtype=torch.float, device=true_X.device)
+    row_X = torch.zeros(true_X.size(-1), dtype=true_X.dtype, device=true_X.device)
     row_X[0] = 1.
-    row_E = torch.zeros(true_E.size(-1), dtype=torch.float, device=true_E.device)
+    row_E = torch.zeros(true_E.size(-1), dtype=true_E.dtype, device=true_E.device)
     row_E[0] = 1.
 
     diag_mask = ~torch.eye(node_mask.size(1), device=node_mask.device, dtype=torch.bool).unsqueeze(0)
